+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
 ## 1.15.1
-### 11-Jun-2024
-- 
+### 29-Jul-2024
+- Added proof-of-concept validations for swiftDialog `2.5.1`'s "blurscreen" control
 
 ## 1.15.0
 ### 11-Jun-2024